refactor: 메뉴 관련 로직 별도의 도메인에서 관리 외
- 특정 메뉴 조회 구현
