+++ src/main/java/com/ll/commars/domain/restaurant/menu/controller/MenuController.java
@@ -1,7 +1,8 @@
-package com.ll.commars.domain.restaurant.controller;
+package com.ll.commars.domain.restaurant.menu.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -9,11 +10,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ll.commars.domain.restaurant.dto.MenuCreateRequest;
-import com.ll.commars.domain.restaurant.dto.MenuCreateResponse;
-import com.ll.commars.domain.restaurant.dto.MenuUpdateRequest;
-import com.ll.commars.domain.restaurant.dto.MenuUpdateResponse;
-import com.ll.commars.domain.restaurant.service.RestaurantMenuService;
+import com.ll.commars.domain.restaurant.menu.dto.MenuCreateRequest;
+import com.ll.commars.domain.restaurant.menu.dto.MenuCreateResponse;
+import com.ll.commars.domain.restaurant.menu.dto.MenuFindResponse;
+import com.ll.commars.domain.restaurant.menu.dto.MenuUpdateRequest;
+import com.ll.commars.domain.restaurant.menu.dto.MenuUpdateResponse;
+import com.ll.commars.domain.restaurant.menu.service.RestaurantMenuService;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -31,6 +33,15 @@ public class MenuController {
 		@RequestBody @Valid MenuCreateRequest request
 	) {
 		MenuCreateResponse response = restaurantMenuService.createMenu(restaurantId, request);
+
+		return ResponseEntity.ok().body(response);
+	}
+
+	@GetMapping("/{menu-id}")
+	public ResponseEntity<MenuFindResponse> getMenu(
+		@PathVariable("menu-id") Long menuId
+	) {
+		MenuFindResponse response = restaurantMenuService.getMenu(menuId);
 
 		return ResponseEntity.ok().body(response);
 	}
