+++ src/main/java/com/ll/commars/domain/restaurant/menu/dto/MenuCreateResponse.java
@@ -1,4 +1,4 @@
-package com.ll.commars.domain.restaurant.dto;
+package com.ll.commars.domain.restaurant.menu.dto;
 
 import com.ll.commars.domain.restaurant.entity.Menu;
