+++ src/main/java/com/ll/commars/domain/restaurant/menu/dto/MenuUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.ll.commars.domain.restaurant.dto;
+package com.ll.commars.domain.restaurant.menu.dto;
 
 public record MenuUpdateRequest(
 	String menuName,
